refatoração no codigo e alteração no parametro das funções de int para double

+++ Calculadora/Operadores.cs
@@ -1,23 +1,21 @@
-﻿
-
-partial class Program
+﻿partial class Program
 {
-    static double soma(int x, int y)
+    static double soma(double x, double y)
     {
         return x + y;
     }
 
-    static double subtracao(int x , int y)
+    static double subtracao(double x , double y)
     {
         return x - y;
     }
 
-    static double multiplicacao(int x , int y)
+    static double multiplicacao(double x , double y)
     {
         return x * y;
     }
 
-    static double divisao(int x , int y)
+    static double divisao(double x , double y)
     {
         return x / y;
     }
